changed vector to float and fixed dependencies

+++ src/Vector.cpp
@@ -5,7 +5,7 @@ Vector::Vector() {
     y = 0;
 }
 
-Vector::Vector(int x, int y) {
+Vector::Vector(float x, float y) {
     this->x = x;
     this->y = y;
 }
@@ -18,11 +18,11 @@ Vector Vector::subtract(Vector b) {
     return Vector(this->x - b.x, this->y - b.y);
 }
 
-Vector Vector::scale(int n) {
+Vector Vector::scale(float n) {
     return Vector(this->x*n, this->y*n);
 }
 
-int Vector::length() {
+float Vector::length() {
     return sqrt((this->x * this->x) + (this->y * this->y));
 }
 
@@ -30,7 +30,7 @@ int Vector::length() {
 //     return this->scale(1/this->length());
 // }
 
-int Vector::dotProduct(Vector b) {
+float Vector::dotProduct(Vector b) {
     return this->x * b.x + this->y * b.y;
 }
 
@@ -42,23 +42,23 @@ bool Vector::operator==(Vector b) {
     return this->equals(b);
 }
 // Getter and Setter
-void Vector::set(int x, int y) {
+void Vector::set(float x, float y) {
     this->x = x;
     this->y = y;
 }
 
-void Vector::setX(int x) {
+void Vector::setX(float x) {
     this->x = x;
 }
 
-void Vector::setY(int y) {
+void Vector::setY(float y) {
     this->y = y;
 }
 
-int Vector::getX() {
+float Vector::getX() {
     return x;
 }
 
-int Vector::getY() {
+float Vector::getY() {
     return y;
 }
